Remove HeapPersistence
