+++ targets/os/win32/nanoCLR.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
-VisualStudioVersion = 15.0.27130.2003
+VisualStudioVersion = 15.0.27130.2027
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "nanoCLR", "nanoCLR\nanoCLR.vcxproj", "{B5515C8A-B056-4320-8EA4-8CACE21A7D2E}"
 	ProjectSection(ProjectDependencies) = postProject
@@ -24,8 +24,6 @@ Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "WireProtocol", "..\..\..\sr
 EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Hardware", "..\..\..\src\CLR\Core\Hardware\Hardware.vcxproj", "{5D50772E-DAC8-407D-972F-37DEB4450B72}"
 EndProject
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "HeapPersistence", "..\..\..\src\CLR\Core\HeapPersistence\HeapPersistence.vcxproj", "{BC2245B0-32E0-410E-B3F2-E2BFA0B503B3}"
-EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "InterruptHandler", "..\..\..\src\CLR\Core\InterruptHandler\InterruptHandler.vcxproj", "{D32EE397-7E49-452D-AC95-AFCB3C962E3D}"
 EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "RPC", "..\..\..\src\CLR\Core\RPC\RPC.vcxproj", "{ED4C887A-EC09-465B-B5D9-075450ADC52D}"
@@ -39,8 +37,6 @@ EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Debugger_stub", "..\..\..\src\CLR\Debugger\Debugger_stub.vcxproj", "{6DABCBF1-245A-4E76-BD36-8BC300C7DD82}"
 EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Diagnostics_stub", "..\..\..\src\CLR\Diagnostics\Diagnostics_stub.vcxproj", "{C97A9845-02DC-46E4-B219-BF5EA8FB28E3}"
-EndProject
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "HeapPersistence_stub", "..\..\..\src\CLR\Core\HeapPersistence\HeapPersistence_stub.vcxproj", "{4646F661-9505-45C2-9573-8409A9DD1149}"
 EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "InterruptHandler_stub", "..\..\..\src\CLR\Core\InterruptHandler\InterruptHandler_stub.vcxproj", "{5772D362-7277-49CF-9B49-42335A3154BD}"
 EndProject
@@ -159,16 +155,6 @@ Global
 		{5D50772E-DAC8-407D-972F-37DEB4450B72}.Release|x64.Build.0 = Release|x64
 		{5D50772E-DAC8-407D-972F-37DEB4450B72}.Release|x86.ActiveCfg = Release|Win32
 		{5D50772E-DAC8-407D-972F-37DEB4450B72}.Release|x86.Build.0 = Release|Win32
-		{BC2245B0-32E0-410E-B3F2-E2BFA0B503B3}.Debug|Any CPU.ActiveCfg = Debug|Win32
-		{BC2245B0-32E0-410E-B3F2-E2BFA0B503B3}.Debug|x64.ActiveCfg = Debug|x64
-		{BC2245B0-32E0-410E-B3F2-E2BFA0B503B3}.Debug|x64.Build.0 = Debug|x64
-		{BC2245B0-32E0-410E-B3F2-E2BFA0B503B3}.Debug|x86.ActiveCfg = Debug|Win32
-		{BC2245B0-32E0-410E-B3F2-E2BFA0B503B3}.Debug|x86.Build.0 = Debug|Win32
-		{BC2245B0-32E0-410E-B3F2-E2BFA0B503B3}.Release|Any CPU.ActiveCfg = Release|Win32
-		{BC2245B0-32E0-410E-B3F2-E2BFA0B503B3}.Release|x64.ActiveCfg = Release|x64
-		{BC2245B0-32E0-410E-B3F2-E2BFA0B503B3}.Release|x64.Build.0 = Release|x64
-		{BC2245B0-32E0-410E-B3F2-E2BFA0B503B3}.Release|x86.ActiveCfg = Release|Win32
-		{BC2245B0-32E0-410E-B3F2-E2BFA0B503B3}.Release|x86.Build.0 = Release|Win32
 		{D32EE397-7E49-452D-AC95-AFCB3C962E3D}.Debug|Any CPU.ActiveCfg = Debug|Win32
 		{D32EE397-7E49-452D-AC95-AFCB3C962E3D}.Debug|x64.ActiveCfg = Debug|x64
 		{D32EE397-7E49-452D-AC95-AFCB3C962E3D}.Debug|x64.Build.0 = Debug|x64
@@ -219,16 +205,6 @@ Global
 		{C97A9845-02DC-46E4-B219-BF5EA8FB28E3}.Release|x64.Build.0 = Release|x64
 		{C97A9845-02DC-46E4-B219-BF5EA8FB28E3}.Release|x86.ActiveCfg = Release|Win32
 		{C97A9845-02DC-46E4-B219-BF5EA8FB28E3}.Release|x86.Build.0 = Release|Win32
-		{4646F661-9505-45C2-9573-8409A9DD1149}.Debug|Any CPU.ActiveCfg = Debug|Win32
-		{4646F661-9505-45C2-9573-8409A9DD1149}.Debug|x64.ActiveCfg = Debug|x64
-		{4646F661-9505-45C2-9573-8409A9DD1149}.Debug|x64.Build.0 = Debug|x64
-		{4646F661-9505-45C2-9573-8409A9DD1149}.Debug|x86.ActiveCfg = Debug|Win32
-		{4646F661-9505-45C2-9573-8409A9DD1149}.Debug|x86.Build.0 = Debug|Win32
-		{4646F661-9505-45C2-9573-8409A9DD1149}.Release|Any CPU.ActiveCfg = Release|Win32
-		{4646F661-9505-45C2-9573-8409A9DD1149}.Release|x64.ActiveCfg = Release|x64
-		{4646F661-9505-45C2-9573-8409A9DD1149}.Release|x64.Build.0 = Release|x64
-		{4646F661-9505-45C2-9573-8409A9DD1149}.Release|x86.ActiveCfg = Release|Win32
-		{4646F661-9505-45C2-9573-8409A9DD1149}.Release|x86.Build.0 = Release|Win32
 		{5772D362-7277-49CF-9B49-42335A3154BD}.Debug|Any CPU.ActiveCfg = Debug|Win32
 		{5772D362-7277-49CF-9B49-42335A3154BD}.Debug|x64.ActiveCfg = Debug|x64
 		{5772D362-7277-49CF-9B49-42335A3154BD}.Debug|x64.Build.0 = Debug|x64
@@ -300,13 +276,11 @@ Global
 		{DD397EC4-844B-4B59-A67B-CAD6B46F1F00} = {76169580-0DE3-48D0-9545-8FC438C54094}
 		{2337C4A4-3AB4-4014-9C29-9EBE32F4510E} = {76169580-0DE3-48D0-9545-8FC438C54094}
 		{5D50772E-DAC8-407D-972F-37DEB4450B72} = {76169580-0DE3-48D0-9545-8FC438C54094}
-		{BC2245B0-32E0-410E-B3F2-E2BFA0B503B3} = {76169580-0DE3-48D0-9545-8FC438C54094}
 		{D32EE397-7E49-452D-AC95-AFCB3C962E3D} = {76169580-0DE3-48D0-9545-8FC438C54094}
 		{ED4C887A-EC09-465B-B5D9-075450ADC52D} = {76169580-0DE3-48D0-9545-8FC438C54094}
 		{16BF6131-A3B8-4B97-B867-E45669FB48BD} = {76169580-0DE3-48D0-9545-8FC438C54094}
 		{6DABCBF1-245A-4E76-BD36-8BC300C7DD82} = {2662E293-2538-40C2-9472-D936B1220EC2}
 		{C97A9845-02DC-46E4-B219-BF5EA8FB28E3} = {2662E293-2538-40C2-9472-D936B1220EC2}
-		{4646F661-9505-45C2-9573-8409A9DD1149} = {2662E293-2538-40C2-9472-D936B1220EC2}
 		{5772D362-7277-49CF-9B49-42335A3154BD} = {2662E293-2538-40C2-9472-D936B1220EC2}
 		{5406EB0C-2243-47E1-9F21-F2D7CA277092} = {2662E293-2538-40C2-9472-D936B1220EC2}
 		{8ACD05E8-42FB-4A49-9D8B-79239F0EED25} = {2662E293-2538-40C2-9472-D936B1220EC2}
