+++ targets/os/win32/nanoCLR/main.cpp
@@ -20,7 +20,6 @@
 #pragma comment(lib, "Debugger.lib")
 #pragma comment(lib, "Diagnostics.lib")
 #pragma comment(lib, "Hardware.lib")
-#pragma comment(lib, "HeapPersistence.lib")
 #pragma comment(lib, "InterruptHandler.lib")
 #pragma comment(lib, "Messaging.lib")
 #pragma comment(lib, "RPC.lib")
@@ -33,7 +32,6 @@
 #pragma comment(lib, "Debugger_stub.lib")
 #pragma comment(lib, "Diagnostics_stub.lib")
 #pragma comment(lib, "Hardware_stub.lib")
-#pragma comment(lib, "HeapPersistence_stub.lib")
 #pragma comment(lib, "InterruptHandler_stub.lib")
 #pragma comment(lib, "IOPort_stub.lib")
 #pragma comment(lib, "Messaging_stub.lib")
